Fix Remote Explorer when moved (#194782)
Part of #194606
